restructuring the api folder
