+++ README.md
@@ -1,1 +1,3 @@
 # orca
+
+Run `python -m api.main` to start the api.
