+++ config/templates/static.py
@@ -1,0 +1,10 @@
+static_dockerfile = """FROM nginx:alpine
+
+WORKDIR /usr/share/nginx/html
+
+COPY . .
+
+EXPOSE 80
+
+CMD ["nginx", "-g", "daemon off;"]
+"""
